# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.BankAccountResponse:
    properties:
      account_name:
        example: PT SIKOn Konveksi
        type: string
      account_number:
        example: "1234567890"
        type: string
      bank_name:
        example: BCA
        type: string
      created_at:
        example: "2023-10-01T15:00:00Z"
        type: string
      id:
        example: 777e4567-e89b-12d3-a456-426614174000
        type: string
      updated_at:
        example: "2023-10-01T15:00:00Z"
        type: string
      user:
        allOf:
        - $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.UserResponse'
        description: Relasi ke User
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.BatchPOResponse:
    properties:
      created_at:
        example: "2026-05-25T10:00:00Z"
        type: string
      end_date:
        example: "2026-06-06T00:00:00Z"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      name:
        example: PO 1 Juni 2026
        type: string
      quota:
        example: 100
        type: integer
      start_date:
        example: "2026-05-30T00:00:00Z"
        type: string
      status:
        example: draft
        type: string
      target_month:
        example: 6
        type: integer
      target_year:
        example: 2026
        type: integer
      updated_at:
        example: "2026-05-25T10:00:00Z"
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.CategoryResponse:
    properties:
      created_at:
        example: "2023-10-01T15:00:00Z"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      image_url:
        example: https://example.com/cat.jpg
        type: string
      name:
        example: Kaos Polos
        type: string
      updated_at:
        example: "2023-10-01T15:00:00Z"
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.CustomerResponse:
    properties:
      address:
        example: Jl. Sudirman No. 123, Jakarta
        type: string
      created_at:
        example: "2023-10-01T15:00:00Z"
        type: string
      created_by:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      creator:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.UserResponse'
      id:
        example: 333e4567-e89b-12d3-a456-426614174000
        type: string
      name:
        example: PT Maju Jaya
        type: string
      phone:
        example: "081234567890"
        type: string
      sales:
        allOf:
        - $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.UserResponse'
        description: 'TAMBAHAN: Munculkan detail User(Sales)'
      sales_id:
        description: TAMBAHAN
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      updated_at:
        example: "2023-10-01T15:00:00Z"
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.OrderItemResponse:
    properties:
      custom_name:
        example: Custom Name
        type: string
      details: {}
      id:
        example: item-uuid
        type: string
      price:
        example: 35000
        type: number
      product:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductResponse'
      product_id:
        example: 999e4567-e89b-12d3-a456-426614174000
        type: string
      qty:
        example: 100
        type: integer
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.OrderResponse:
    properties:
      batch_po:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.BatchPOResponse'
      batch_po_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      courier_name:
        example: JNE Trucking
        type: string
      created_at:
        example: "2023-10-01T15:00:00Z"
        type: string
      customer:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.CustomerResponse'
      customer_id:
        example: 333e4567-e89b-12d3-a456-426614174000
        type: string
      discount_amount:
        example: 0
        type: number
      id:
        example: ord-uuid
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.OrderItemResponse'
        type: array
      notes:
        example: Tolong packing kayu
        type: string
      order_number:
        example: ORD-20231001-1234
        type: string
      order_status:
        example: pending
        type: string
      payment_status:
        example: unpaid
        type: string
      sales:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.UserResponse'
      sales_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      shipping_address:
        example: Jl. Sudirman No. 123, Jakarta
        type: string
      shipping_cost:
        example: 50000
        type: number
      subtotal:
        example: 3500000
        type: number
      tax_ph:
        example: 0
        type: number
      tax_ppn:
        example: 0
        type: number
      terms_conditions:
        example: DP Minimal 50%
        type: string
      total_amount:
        example: 3550000
        type: number
      updated_at:
        example: "2023-10-01T15:00:00Z"
        type: string
      valid_until:
        example: "2026-06-15T00:00:00Z"
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.PaymentCreateRequest:
    properties:
      amount:
        example: 1000000
        type: number
      bank_account_id:
        example: 777e4567-e89b-12d3-a456-426614174000
        type: string
      order_id:
        example: ord-uuid
        type: string
      payment_date:
        example: "2026-10-02T10:00:00Z"
        type: string
      payment_type:
        enum:
        - dp
        - settlement
        - installment
        example: dp
        type: string
      reference_number:
        example: TRX-0987654321
        type: string
    required:
    - amount
    - bank_account_id
    - order_id
    - payment_type
    - reference_number
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.PaymentResponse:
    properties:
      amount:
        example: 1000000
        type: number
      bank_account:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.BankAccountResponse'
      bank_account_id:
        example: 777e4567-e89b-12d3-a456-426614174000
        type: string
      created_at:
        example: "2026-10-02T10:05:00Z"
        type: string
      id:
        example: pay-uuid
        type: string
      order:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.OrderResponse'
      order_id:
        example: ord-uuid
        type: string
      payment_date:
        example: "2026-10-02T10:00:00Z"
        type: string
      payment_type:
        example: dp
        type: string
      reference_number:
        example: TRX-0987654321
        type: string
      status:
        description: pending, verified, rejected
        example: pending
        type: string
      updated_at:
        example: "2026-10-02T10:05:00Z"
        type: string
      verified_at:
        example: "2026-10-02T11:00:00Z"
        type: string
      verified_by:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.UserResponse'
      verified_by_id:
        example: user-finance-uuid
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.PaymentUpdateRequest:
    properties:
      payment_type:
        enum:
        - dp
        - settlement
        - installment
        example: settlement
        type: string
      reference_number:
        example: TRX-0987654322-REVISI
        type: string
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.PaymentVerifyRequest:
    properties:
      status:
        enum:
        - verified
        - rejected
        example: verified
        type: string
    required:
    - status
    type: object
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductFabricResponse:
    properties:
      base_price:
        type: number
      care_instruction:
        type: string
      colors:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.FabricColorResponse'
        type: array
      composition:
        type: string
      description:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      price_adjustment:
        type: number
    type: object
  github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductImageResponse:
    properties:
      id:
        type: string
      image_url:
        type: string
      is_primary:
        type: boolean
    type: object
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductModelResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        type: string
      views:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductModelViewResponse'
        type: array
    type: object
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductResponse:
    properties:
      base_price:
        example: 185000
        type: number
      category:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.CategoryResponse'
      category_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      created_at:
        type: string
      description:
        example: Kemeja taktikal bahan ripstop cotton
        type: string
      design_model:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductModelResponse'
      fabric_summary:
        example: Ripstop
        type: string
      fabrics:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductFabricResponse'
        type: array
      gsm_info:
        example: 210gsm
        type: string
      id:
        example: 999e4567-e89b-12d3-a456-426614174000
        type: string
      images:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.ProductImageResponse'
        type: array
      key_features:
        items:
          type: string
        type: array
      name:
        example: Kemeja Taktikal Premium 7200
        type: string
      rating:
        example: 4.9
        type: number
      review_count:
        example: 318
        type: integer
      slug:
        example: kemeja-taktikal-premium-7200
        type: string
      sold_count:
        example: 1240
        type: integer
      updated_at:
        type: string
      wholesale:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.WholesalePriceResponse'
        type: array
    type: object
...
  github_com_faridlan_sikon-api_internal_delivery_http_dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      image_url:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      phone:
        type: string
      role:
        type: string
      sort_order:
        type: integer
      status_text:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_domain.PaginationMeta:
    properties:
      current_page:
        type: integer
      limit:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_faridlan_sikon-api_internal_utils.EmptyObj:
    type: object
  github_com_faridlan_sikon-api_internal_utils.ErrorResponse:
    properties:
      detail:
        type: string
      error:
        example: pesan error
        type: string
    type: object
...
  ? github_com_faridlan_sikon-api_internal_utils.PaginatedResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_PaymentResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.PaymentResponse'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_domain.PaginationMeta'
    type: object
...
  ? github_com_faridlan_sikon-api_internal_utils.SuccessResponse-array_github_com_faridlan_sikon-api_internal_delivery_http_dto_PaymentResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.PaymentResponse'
        type: array
      message:
        type: string
    type: object
...
  ? github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_PaymentResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.PaymentResponse'
      message:
        type: string
    type: object
...
  github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_utils_EmptyObj:
    properties:
      data:
        $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.EmptyObj'
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Ini adalah dokumentasi API untuk MVP ERP SIKOn.
  title: SIKOn API (Sistem Integrasi Konveksi Online)
  version: "1.0"
paths:
...
  /payments:
    get:
      description: 'Mengambil daftar seluruh histori pembayaran dengan fitur pagination
        dan filter (termasuk filter status verifikasi: pending, verified, rejected).'
      parameters:
      - default: 1
        description: Nomor Halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Batas Data per Halaman
        in: query
        name: limit
        type: integer
      - description: Cari No Referensi atau Order ID
        in: query
        name: search
        type: string
      - description: Filter Tipe Pembayaran (dp, settlement, installment)
        in: query
        name: payment_type
        type: string
      - description: Filter Status Verifikasi (pending, verified, rejected)
        in: query
        name: status
        type: string
      - description: Tanggal Mulai Pembayaran (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Tanggal Akhir Pembayaran (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.PaginatedResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_PaymentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Payments
      tags:
      - Payments
    post:
      consumes:
      - application/json
      description: Memproses pembayaran (DP/Lunas) baru oleh Sales/Admin. Pembayaran
        yang baru di-submit akan berstatus `pending` sampai diverifikasi oleh Divisi
        Finance.
      parameters:
      - description: Data Pembayaran
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.PaymentCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_PaymentResponse'
        "400":
          description: Nominal bayar lebih besar dari sisa tagihan
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "404":
          description: Order atau Bank tidak ditemukan
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "409":
          description: Pesanan sudah lunas sepenuhnya
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Process Payment
      tags:
      - Payments
  /payments/{id}:
    delete:
      description: Menghapus data/histori pembayaran secara permanen.
      parameters:
      - description: Payment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_utils_EmptyObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Payment
      tags:
      - Payments
    get:
      description: Mengambil detail data pembayaran berdasarkan ID pembayaran
      parameters:
      - description: Payment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Payment
      tags:
      - Payments
    put:
      consumes:
      - application/json
      description: Memperbarui data meta pembayaran seperti nomor referensi atau tipe
        pembayaran.
      parameters:
      - description: Payment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data Update Pembayaran
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.PaymentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Payment
      tags:
      - Payments
  /payments/{id}/verify:
    patch:
      consumes:
      - application/json
      description: Memverifikasi atau menolak bukti pembayaran. <br><br> **EFEK SAMPING
        OTOMATIS:** <br> 1. Jika di-`verified`, sistem menghitung uang sah dan mengubah
        `payment_status` Order ke `partial` atau `paid`. <br> 2. Uang baru resmi masuk
        ke Kas / Laporan Akuntansi.
      parameters:
      - description: Payment ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Status Verifikasi (verified / rejected)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_faridlan_sikon-api_internal_delivery_http_dto.PaymentVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-github_com_faridlan_sikon-api_internal_delivery_http_dto_PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Verify Payment (Finance Approval)
      tags:
      - Payments
  /payments/order/{order_id}:
    get:
      description: Mengambil rincian histori/cicilan pembayaran khusus untuk satu
        pesanan tertentu.
      parameters:
      - description: Order ID (UUID)
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.SuccessResponse-array_github_com_faridlan_sikon-api_internal_delivery_http_dto_PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_faridlan_sikon-api_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Payments by Order ID
      tags:
      - Payments
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"