# internal/delivery/http/payment_handler.go
package http

import (
	"github.com/gofiber/fiber/v2"

	"github.com/faridlan/sikon-api/internal/delivery/http/dto"
	"github.com/faridlan/sikon-api/internal/domain"
	"github.com/faridlan/sikon-api/internal/utils"
)

type PaymentHandler interface {
	ProcessPayment(c *fiber.Ctx) error
	GetPayment(c *fiber.Ctx) error
	ListPayments(c *fiber.Ctx) error
	UpdatePayment(c *fiber.Ctx) error
	DeletePayment(c *fiber.Ctx) error
	GetPaymentsByOrderID(c *fiber.Ctx) error
	VerifyPayment(c *fiber.Ctx) error // BARU
}

type paymentHandler struct {
	paymentUsecase domain.PaymentUsecase
}

func NewPaymentHandler(pu domain.PaymentUsecase) PaymentHandler {
	return &paymentHandler{
		paymentUsecase: pu,
	}
}

// @Summary Process Payment
// @Description Memproses pembayaran (DP/Lunas) baru oleh Sales/Admin. Pembayaran yang baru di-submit akan berstatus `pending` sampai diverifikasi oleh Divisi Finance.
// @Tags Payments
// @Accept json
// @Produce json
// @Param request body dto.PaymentCreateRequest true "Data Pembayaran"
// @Success 201 {object} utils.SuccessResponse[dto.PaymentResponse]
// @Failure 400 {object} utils.ErrorResponse "Nominal bayar lebih besar dari sisa tagihan"
// @Failure 404 {object} utils.ErrorResponse "Order atau Bank tidak ditemukan"
// @Failure 409 {object} utils.ErrorResponse "Pesanan sudah lunas sepenuhnya"
// @Failure 500 {object} utils.ErrorResponse
// @Failure 401 {object} utils.ErrorResponse "Unauthorized"
// @Security BearerAuth
// @Router /payments [post]
func (h *paymentHandler) ProcessPayment(c *fiber.Ctx) error {
	var req dto.PaymentCreateRequest
	if err := c.BodyParser(&req); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, "Gagal memparsing request body")
	}
	if err := utils.ValidateStruct(&req); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, err.Error())
	}

	domainReq := domain.PaymentCreateInput{
		OrderID:         req.OrderID,
		BankAccountID:   req.BankAccountID,
		Amount:          req.Amount,
		PaymentDate:     req.PaymentDate,
		ReferenceNumber: req.ReferenceNumber,
		PaymentType:     domain.PaymentType(req.PaymentType),
	}

	payment, err := h.paymentUsecase.ProcessPayment(c.Context(), domainReq)
	if err != nil {
		return utils.HandleDomainError(c, err)
	}

	return utils.SendSuccess(c, fiber.StatusCreated, "Berhasil memproses pembayaran", dto.ToPaymentResponse(payment))
}

// @Summary Get Payment
// @Description Mengambil detail data pembayaran berdasarkan ID pembayaran
// @Tags Payments
// @Produce json
// @Param id path string true "Payment ID (UUID)"
// @Success 200 {object} utils.SuccessResponse[dto.PaymentResponse]
// @Failure 400 {object} utils.ErrorResponse
// @Failure 404 {object} utils.ErrorResponse
// @Failure 500 {object} utils.ErrorResponse
// @Failure 401 {object} utils.ErrorResponse "Unauthorized"
// @Security BearerAuth
// @Router /payments/{id} [get]
func (h *paymentHandler) GetPayment(c *fiber.Ctx) error {
	id := c.Params("id")
	if err := utils.ValidateUUID(id, "id"); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, err.Error())
	}

	payment, err := h.paymentUsecase.GetPayment(c.Context(), id)
	if err != nil {
		return utils.HandleDomainError(c, err)
	}

	return utils.SendSuccess(c, fiber.StatusOK, "Berhasil mengambil data pembayaran", dto.ToPaymentResponse(payment))
}

// @Summary List Payments
// @Description Mengambil daftar seluruh histori pembayaran dengan fitur pagination dan filter (termasuk filter status verifikasi: pending, verified, rejected).
// @Tags Payments
// @Produce json
// @Param page query int false "Nomor Halaman" default(1)
// @Param limit query int false "Batas Data per Halaman" default(10)
// @Param search query string false "Cari No Referensi atau Order ID"
// @Param payment_type query string false "Filter Tipe Pembayaran (dp, settlement, installment)"
// @Param status query string false "Filter Status Verifikasi (pending, verified, rejected)"
// @Param start_date query string false "Tanggal Mulai Pembayaran (YYYY-MM-DD)"
// @Param end_date query string false "Tanggal Akhir Pembayaran (YYYY-MM-DD)"
// @Success 200 {object} utils.PaginatedResponse[dto.PaymentResponse]
// @Failure 500 {object} utils.ErrorResponse
// @Failure 401 {object} utils.ErrorResponse "Unauthorized"
// @Security BearerAuth
// @Router /payments [get]
func (h *paymentHandler) ListPayments(c *fiber.Ctx) error {
	page := c.QueryInt("page", 1)
	limit := c.QueryInt("limit", 10)

	filter := domain.PaymentFilter{
		Search:      c.Query("search"),
		PaymentType: c.Query("payment_type"),
		Status:      c.Query("status"), // BARU: Filter berdasarkan status verifikasi
		StartDate:   c.Query("start_date"),
		EndDate:     c.Query("end_date"),
	}

	query := domain.PaginationQuery{Page: page, Limit: limit}

	payments, meta, err := h.paymentUsecase.ListPayments(c.Context(), query, filter)
	if err != nil {
		return utils.HandleDomainError(c, err)
	}

	return utils.SendSuccessPaginated(c, "Berhasil mengambil daftar pembayaran", dto.ToPaymentResponseList(payments), meta)
}

// @Summary Update Payment
// @Description Memperbarui data meta pembayaran seperti nomor referensi atau tipe pembayaran.
// @Tags Payments
// @Accept json
// @Produce json
// @Param id path string true "Payment ID (UUID)"
// @Param request body dto.PaymentUpdateRequest true "Data Update Pembayaran"
// @Success 200 {object} utils.SuccessResponse[dto.PaymentResponse]
// @Failure 400 {object} utils.ErrorResponse
// @Failure 404 {object} utils.ErrorResponse
// @Failure 500 {object} utils.ErrorResponse
// @Failure 401 {object} utils.ErrorResponse "Unauthorized"
// @Security BearerAuth
// @Router /payments/{id} [put]
func (h *paymentHandler) UpdatePayment(c *fiber.Ctx) error {
	id := c.Params("id")
	if err := utils.ValidateUUID(id, "id"); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, err.Error())
	}

	var req dto.PaymentUpdateRequest
	if err := c.BodyParser(&req); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, "Gagal memparsing request body")
	}
	if err := utils.ValidateStruct(&req); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, err.Error())
	}

	domainReq := domain.PaymentUpdateInput{
		ReferenceNumber: req.ReferenceNumber,
		PaymentType:     domain.PaymentType(req.PaymentType),
	}

	payment, err := h.paymentUsecase.UpdatePayment(c.Context(), id, domainReq)
	if err != nil {
		return utils.HandleDomainError(c, err)
	}

	return utils.SendSuccess(c, fiber.StatusOK, "Berhasil memperbarui data pembayaran", dto.ToPaymentResponse(payment))
}

// @Summary Verify Payment (Finance Approval)
// @Description Memverifikasi atau menolak bukti pembayaran. <br><br> **EFEK SAMPING OTOMATIS:** <br> 1. Jika di-`verified`, sistem menghitung uang sah dan mengubah `payment_status` Order ke `partial` atau `paid`. <br> 2. Uang baru resmi masuk ke Kas / Laporan Akuntansi.
// @Tags Payments
// @Accept json
// @Produce json
// @Param id path string true "Payment ID (UUID)"
// @Param request body dto.PaymentVerifyRequest true "Status Verifikasi (verified / rejected)"
// @Success 200 {object} utils.SuccessResponse[dto.PaymentResponse]
// @Failure 400 {object} utils.ErrorResponse
// @Failure 404 {object} utils.ErrorResponse
// @Failure 500 {object} utils.ErrorResponse
// @Failure 401 {object} utils.ErrorResponse "Unauthorized"
// @Security BearerAuth
// @Router /payments/{id}/verify [patch]
func (h *paymentHandler) VerifyPayment(c *fiber.Ctx) error {
	id := c.Params("id")
	if err := utils.ValidateUUID(id, "id"); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, err.Error())
	}

	var req dto.PaymentVerifyRequest
	if err := c.BodyParser(&req); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, "Gagal memparsing request body")
	}
	if err := utils.ValidateStruct(&req); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, err.Error())
	}

	userID := c.Get("X-User-Id")

	// Validasi apakah X-User-Id benar-benar format UUID
	if err := utils.ValidateUUID(userID, "x-user-id"); err != nil {
		// Jika X-User-Id dari FE berupa dummy string biasa (bukan UUID format),
		// kosongkan nilainya agar diset NULL atau fallback ke nilai aman tanpa merusak SQL constraint
		userID = ""
	}

	input := domain.PaymentVerifyInput{
		Status:       domain.PaymentVerificationStatus(req.Status),
		VerifiedByID: userID,
	}

	payment, err := h.paymentUsecase.VerifyPayment(c.Context(), id, input)
	if err != nil {
		return utils.HandleDomainError(c, err)
	}

	return utils.SendSuccess(c, fiber.StatusOK, "Berhasil memperbarui status verifikasi pembayaran", dto.ToPaymentResponse(payment))
}

// @Summary Delete Payment
// @Description Menghapus data/histori pembayaran secara permanen.
// @Tags Payments
// @Produce json
// @Param id path string true "Payment ID (UUID)"
// @Success 200 {object} utils.SuccessResponse[utils.EmptyObj]
// @Failure 400 {object} utils.ErrorResponse
// @Failure 404 {object} utils.ErrorResponse
// @Failure 500 {object} utils.ErrorResponse
// @Failure 401 {object} utils.ErrorResponse "Unauthorized"
// @Security BearerAuth
// @Router /payments/{id} [delete]
func (h *paymentHandler) DeletePayment(c *fiber.Ctx) error {
	id := c.Params("id")
	if err := utils.ValidateUUID(id, "id"); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, err.Error())
	}

	if err := h.paymentUsecase.DeletePayment(c.Context(), id); err != nil {
		return utils.HandleDomainError(c, err)
	}

	return utils.SendSuccess(c, fiber.StatusOK, "Berhasil menghapus pembayaran", nil)
}

// @Summary Get Payments by Order ID
// @Description Mengambil rincian histori/cicilan pembayaran khusus untuk satu pesanan tertentu.
// @Tags Payments
// @Produce json
// @Param order_id path string true "Order ID (UUID)"
// @Success 200 {object} utils.SuccessResponse[[]dto.PaymentResponse]
// @Failure 400 {object} utils.ErrorResponse
// @Failure 404 {object} utils.ErrorResponse
// @Failure 500 {object} utils.ErrorResponse
// @Failure 401 {object} utils.ErrorResponse "Unauthorized"
// @Security BearerAuth
// @Router /payments/order/{order_id} [get]
func (h *paymentHandler) GetPaymentsByOrderID(c *fiber.Ctx) error {
	orderID := c.Params("order_id")
	if err := utils.ValidateUUID(orderID, "order_id"); err != nil {
		return utils.SendError(c, fiber.StatusBadRequest, err.Error())
	}

	payments, err := h.paymentUsecase.GetPaymentsByOrderID(c.Context(), orderID)
	if err != nil {
		return utils.HandleDomainError(c, err)
	}

	return utils.SendSuccess(c, fiber.StatusOK, "Berhasil mengambil daftar pembayaran untuk pesanan", dto.ToPaymentResponseList(payments))
}